requirements.txt
certifi==2019.11.28
Click==7.0
dataclasses==0.6
Django==2.2
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
importlib-metadata==4.8.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
MarkupSafe==1.1.1
numpy==1.17.4
pandas==1.1.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
typing-extensions==3.10.0.2
virtualenv==16.7.8
Werkzeug==0.16.0
wincertstore==0.2
zipp==3.5.0
astrapy==0.2.1
cassandra-driver==3.25.0